requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.6
django-bootstrap4==2.3.1
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2021.1
soupsieve==2.2
sqlparse==0.4.1
whitenoise==5.2.0
django-embed-video==1.4.0
